backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pymongo
speechrecognition
pydub
pyjwt
bcrypt
faster-whisper
langdetect
numpy
scipy
sounddevice
torch
